Moves the creation of the arrangeGrob inside the png() call.
This resolves the incredibly annoying creation of the Rplots.pdf file
that seems to be caused by an internal call during arrangeGrob to the
dimensions of the display device. The resulting code is quite
inelegant, but will have to do until this side effect is resolved.

As a bonus, it prevents R from throwing the 15 warnings related to
missing fonts.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>